ability to pull a root record, useful for sharing
